move mock config from testutil to configs directory
