+++ pkg/gofr/gofr_test.go
@@ -224,7 +224,7 @@ func TestEnableBasicAuthWithFunc(t *testing.T) {
 		w.WriteHeader(http.StatusOK)
 	}))
 
-	c := container.NewContainer(testutil.NewMockConfig(nil))
+	c := container.NewContainer(config.NewMockConfig(nil))
 
 	// Initialize a new App instance
 	a := &App{
@@ -291,13 +291,13 @@ func Test_AddRESTHandlers(t *testing.T) {
 }
 
 func Test_initTracer(t *testing.T) {
-	mockConfig1 := testutil.NewMockConfig(map[string]string{
+	mockConfig1 := config.NewMockConfig(map[string]string{
 		"TRACE_EXPORTER": "zipkin",
 		"TRACER_HOST":    "localhost",
 		"TRACER_PORT":    "2005",
 	})
 
-	mockConfig2 := testutil.NewMockConfig(map[string]string{
+	mockConfig2 := config.NewMockConfig(map[string]string{
 		"TRACE_EXPORTER": "jaeger",
 		"TRACER_HOST":    "localhost",
 		"TRACER_PORT":    "2005",
@@ -329,7 +329,7 @@ func Test_initTracer(t *testing.T) {
 }
 
 func Test_initTracer_invalidConfig(t *testing.T) {
-	mockConfig := testutil.NewMockConfig(map[string]string{
+	mockConfig := config.NewMockConfig(map[string]string{
 		"TRACE_EXPORTER": "abc",
 		"TRACER_HOST":    "localhost",
 		"TRACER_PORT":    "2005",
